# Faster range