After
```
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_10_percent                                  ... bench:     112,407 ns/iter (+/- 8,901)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_10_percent_intersect_with_10_percent        ... bench:     284,118 ns/iter (+/- 14,614)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_10_percent_intersect_with_10_percent_multi  ... bench:     870,234 ns/iter (+/- 38,734)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_10_percent_intersect_with_90_percent        ... bench:     370,674 ns/iter (+/- 16,682)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_10_percent_intersect_with_90_percent_multi  ... bench:     851,692 ns/iter (+/- 73,837)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_10_percent_multi                            ... bench:     622,292 ns/iter (+/- 21,351)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_1_percent                                   ... bench:     100,109 ns/iter (+/- 24,225)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_1_percent_intersect_with_10_percent         ... bench:     179,461 ns/iter (+/- 12,052)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_1_percent_intersect_with_10_percent_multi   ... bench:     458,858 ns/iter (+/- 23,595)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_1_percent_intersect_with_1_percent          ... bench:     154,085 ns/iter (+/- 6,997)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_1_percent_intersect_with_1_percent_multi    ... bench:     369,613 ns/iter (+/- 7,035)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_1_percent_intersect_with_90_percent         ... bench:     182,634 ns/iter (+/- 3,470)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_1_percent_intersect_with_90_percent_multi   ... bench:     453,847 ns/iter (+/- 6,500)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_1_percent_multi                             ... bench:     485,043 ns/iter (+/- 54,427)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_90_percent                                  ... bench:     202,176 ns/iter (+/- 34,391)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_90_percent_intersect_with_10_percent        ... bench:     529,532 ns/iter (+/- 78,635)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_90_percent_intersect_with_10_percent_multi  ... bench:   1,897,125 ns/iter (+/- 400,715)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_90_percent_intersect_with_1_percent         ... bench:     195,786 ns/iter (+/- 34,028)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_90_percent_intersect_with_1_percent_multi   ... bench:     970,128 ns/iter (+/- 71,751)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_90_percent_intersect_with_90_percent        ... bench:   1,895,994 ns/iter (+/- 392,031)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_90_percent_intersect_with_90_percent_multi  ... bench:   3,259,080 ns/iter (+/- 833,771)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_90_percent_multi                            ... bench:   1,469,404 ns/iter (+/- 121,789)
```

Before
```
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_10_percent                                  ... bench:     316,482 ns/iter (+/- 10,593)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_10_percent_intersect_with_10_percent        ... bench:     471,190 ns/iter (+/- 15,529)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_10_percent_intersect_with_10_percent_multi  ... bench:   1,083,478 ns/iter (+/- 86,525)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_10_percent_intersect_with_90_percent        ... bench:     538,146 ns/iter (+/- 46,114)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_10_percent_intersect_with_90_percent_multi  ... bench:   1,190,530 ns/iter (+/- 262,962)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_10_percent_multi                            ... bench:     939,202 ns/iter (+/- 58,482)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_1_percent                                   ... bench:     271,097 ns/iter (+/- 6,690)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_1_percent_intersect_with_10_percent         ... bench:     323,999 ns/iter (+/- 7,342)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_1_percent_intersect_with_10_percent_multi   ... bench:     705,554 ns/iter (+/- 39,835)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_1_percent_intersect_with_1_percent          ... bench:     272,626 ns/iter (+/- 10,727)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_1_percent_intersect_with_1_percent_multi    ... bench:     554,363 ns/iter (+/- 32,966)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_1_percent_intersect_with_90_percent         ... bench:     350,479 ns/iter (+/- 16,337)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_1_percent_intersect_with_90_percent_multi   ... bench:     707,452 ns/iter (+/- 23,216)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_1_percent_multi                             ... bench:     714,395 ns/iter (+/- 39,845)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_90_percent                                  ... bench:     459,379 ns/iter (+/- 106,714)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_90_percent_intersect_with_10_percent        ... bench:     713,999 ns/iter (+/- 71,580)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_90_percent_intersect_with_10_percent_multi  ... bench:   2,166,998 ns/iter (+/- 239,663)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_90_percent_intersect_with_1_percent         ... bench:     324,636 ns/iter (+/- 47,800)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_90_percent_intersect_with_1_percent_multi   ... bench:   1,137,768 ns/iter (+/- 82,548)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_90_percent_intersect_with_90_percent        ... bench:   1,849,210 ns/iter (+/- 219,530)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_90_percent_intersect_with_90_percent_multi  ... bench:   3,162,208 ns/iter (+/- 145,310)
test query::range_query::range_query_ip_fastfield::bench::bench_ip_range_hit_90_percent_multi                            ... bench:   1,698,148 ns/iter (+/- 25,987)
```